Moved LocalisedObject & ILocalisedObject into the LocalisedObjects folder
