+++ Editor/LocalisationKeySchema.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using DUCK.Localisation.LocalisedObjects;
 using UnityEngine;
 
 namespace DUCK.Localisation.Editor
