+++ Editor/LocalisedObject.Editor.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using DUCK.Localisation.LocalisedObjects;
 using UnityEditor;
 using UnityEngine;
